Initialize new gem with bundler
This is the skeleton code for a new gem that is auto-generated by
bundler 2.2.3 by running:

    bundle gem --coc --mit --test=rspec rails_cursor_pagination

The rails_cursor_pagination.gemspec file has been filled with the
current details. However, some URLs may be subject to change soon.

Also, added .ruby-version and .ruby-gemset files and adjusted the
default .rubocop.yml configuration a bit.