Add basic setup to allow for pagination
The new RailsCursorPagination::Paginator class is where the actual
pagination will happen. To be able to test this properly, set up
ActiveRecord as a dependency and configure sqlite3 for testing.

We will use an in-memory database to execute the tests against. Set up
a table `posts` with 7 records and a dummy method to return the first 5
records. Add specs accordingly.