Add memoization to prevent multiple DB queries
To prevent us making the same query multiple times, use memoization to
"remember" the return value of a given method via an instance variable.
If the same method is queried again, it will simply return the pre-
calculated value.

This will increase the memory footprint of the gem a bit, but decrease
the load on the database. Since IO is normally the bottleneck of any
application, this should speed up the pagaination.