Set default page size to 10
This is a bit more real-world default page size that better meets the
needs of our app.

By adjusting the spec, we can also see that in case of a custom
`order_by` query, the ID is added to a string and is therefore used for
string sorting (instead of integer sorting). This causes the order to be
a bit counter-intuitive: '1' < '10' < '2'.

But in this case, we only need to ensure a unique order that is
independent of the oder direction. And this is achieved by the ID, no
matter if it is regarded as a string or an integer.