Update README.md with usage details
Add a detailled decsription on what this gem is, how it can be installed
and used, as well as how it works internally.

This is in part a duplication of what we already have as a module
comment in the `rails_cuirsor_pagination.rb` file. But it still makes
sense to repeat it in the README so that everyone can easily find it
when first accessing the gem.