Add compatibility for ActiveRecord version 6
This gem has successfully been tested to work with AR 6.0. Therefore, we
no longer need to lock it to a 5.x version of the gem.

We are currently not locking the upper version of the gem as we expect
next major releases to keep compatibility. But this will need to be
tested on a case-by-case basis once new versions are released.

Also, remove left-over and now deprecated instructions from gemspec
file.