Fix pagination of relations with SELECT query
If the given relation has `#select`-ed some custom columns, the old
logic caused an error on MySQL. This was caused due to:

```ruby
Membership.select(:id, :state).count
```

generating this SQL query:

```sql
SELECT COUNT(id, state) FROM `memberships`
```

which throws this error in MySQL:

```
Mysql2::Error: You have an error in your SQL syntax; check the manual
  that corresponds to your MySQL server version for the right syntax to
  use near ', state) FROM `memberships`' at line 1
```

To prevent this, we can use `ActiveRecord::Relation#size` which instead
generates this SQL query (if the records aren't yet loaded):

```sql
SELECT COUNT(*) FROM `memberships`
```

In the case that the records are already loaded, `#size` just returns
the length of the loaded records.

The tests have been updated to check for this error in all given
combinations of parameters.