Fix pagination for relations without selected ID
The relation passed in by the user can have certain fields selected.
This is common and good practice to reduce the data that is being loaded
from the database. However, for this cursor pagination algorithm to
work, it is crucial that each record has a unique field which can be
used to generate its cursor from. This gem uses the record's ID as that
field.

For this reason, it is paramount that each record has its ID retrieved
from the database. To ensure this, we now manually check if any columns
have been `#select`-ed by the used. If so, and if the `:id` field is not
part of the selected field, we manually add it to the query.

Also, add a section to the README.md file to explain this behavior.