Move repository to public Github and allow publishing to Rubygems
We want to make this gem available to the general public. Therefore,
we migrate the repository to github.com and will start publishing the
gem on the public rubygems gem host.

To make this possible, change the homepage path of the gem metadata and
remove the push host restriction. Also, update the links in README.md.