Fix license to mention XING
This gem was created while working at and for XING GmbH & Co. KG, which
means that the copyright is in the name of the corporation. My personal
name got in here when initially creating the gem from my local machine
and keeping the default, but the intellectual ownership has always been
with XING. Therefore, updating the LICENSE.txt to reflect this.