Add spec to test that new versions are mentioned in CHANGELOG.md
We want to verify that each version of the gem is properly described in
the changelog. This allows users of the gem to better understand what is
changing with each version and what they might need to do to be able to
upgrade.

Also, add a section to the CHANGELOG.md file to mention the upcoming
public release of this gem.