Update gems in Gemfile.lock
There was a vulnerability in activerecord >= 6.1.0 and <= 6.1.2.0.
See more here: https://github.com/advisories/GHSA-8hc4-xxm3-5ppp

In our testing environment we used version 6.1.1 which was affected by
this. Therefore, update this as well as all other dependencies for the
testing enviornment.

The comatibility of the gem with older versions of activerecord will
remain the same.