Improve gemspec file by removing unneeded files
when first creating the gem using bundler. It caused gem releases to
include a lot of files like the content of the `bin` folder or the
Gemfile used for testing that are not relevant for gem users.

Furthermore, it also referenced an executable directory that does not
exist since this gem does not provide any executables.

Therefore, adjust the gemspec file accordingly to define the `files` as
only the bare-minimum required to get the gem working on a clients
machine (along with the relevant documentation and licensing info).