Remove `ORDER BY` clause from `COUNT` queries
Since we just used the pre-assembled relations to then call `#size` on
them, this caused queries like this to be sent to the database:
```
SELECT COUNT(*) FROM (
  SELECT 1 AS one
  FROM `posts`
  ORDER BY `posts`.`author` DESC, `posts`.`id` DESC
  LIMIT 3
) subquery_for_count
```

The `ORDER BY` part will never influence the number of returned records.
But depending on the used database and their query optimization
settings, it will make the query less performant.

Therefore, remove any order from the queries before calling `#size` on
them. With `ActiveRecord` this can be achieved by calling `#reorder` on
the relation with an empty string.