Ensure records are fetched early to avoid multiple DB queries
One of the optimizations that cursor pagination allows is fetching one
more record than required from the database to then be able to
efficiently determine if there is another page (which would contain at
the least this extra item). However, the way this was implemented it
actually triggered individual queries, one to load the records and
another `COUNT` query to determine the amount of records with the
additional one. This was due to how ActiveRecord tries to delay fetching
records of a relation to the latest possible point.

To prevent this, we can call `#fetch` manually to request the records to
be loaded earlier.