Add explanation to README.md about ordering by arbitrary SQL
The gem in its current version does not allow ordering by any arbitrary
SQL statements. Only SQL columns can be used.

This is on one side due to performance reasons, but on the other side
also due to the added high complexity of supporting arbitrary SQL
queries. We would have to ensure that it doesn't open the door for SQL
injection attacks. We would also have to return the value of this query
so that it can be properly encoded into the cursor.

For now, this is out of scope for this gem.