Refactor Cursor into separate class (#74)
The logic around cursors is core to this gem and complex enough that it
warrants its own class along with a good test coverage. This also allows
to create cursors from the "outside" and inject them into the pagination
if the business logic of one's project so requires.

Huge shout-out to @myxoh for taking on this refactoring.