Remove nesting of error classes
From the outside, the user is not interested in which nested class an
error appeared. Instead, the client wants to be able to catch meaningful
errors to then handle them accordingly.

The recent and overdue refactoring of extracting the cursor into its own
class spread the locations from where errors can occur across two
classes. To simplify usage of the gem, this is now consolidated into
two error classes mounted right under the main module that should allow
for a uniform error handling.

This constitutes a breaking change as error handling code within clients
will need to be adjusted.