Add a limit attribute to be used instead of first or last (#73)
Depending on the FE client used it does not always make sense to have a
different parameter to choose the number of records retrieved depending
on whether forward- or backward-pagination is used. Internally, the
`first` and `last` parameters are treated the same. 
This change now
also allows sending a `limit` parameter and the library will handle it
as `first` _or_  `last ` depending on the pagination direction.

Co-authored-by: Nicolas Fricke <mail@nicolasfricke.de>