fix: undefined column error when using custom order field
When providing a relation that selects a virtual field for ordering and does not also select a wildcard, the query will append the order field a second time.

For example:

```ruby
relation = Member.select('members.*', 'random() * 100 as random_order')
paginate(relation, order_by: :random_order)

#=> "PG::UndefinedColumn: ERROR:  column "random_order" does not exist

# Resulting query:
# "SELECT members.*, random() * 100 as random_order, random_order FROM members..."
```

This is most likely to happen on a query that relies on a join, as with a `has_many :through` association, where calling `select('*')` on the association is liable to return unpredictable results due to returning multiple ID columns, and so on.

This fixes the issue by not assuming that potential order columns are defined in the DB schema, and by allowing for wildcard select's that are table specific. (Although I wonder if it would be easier to not bother with this kind of check when the application can be responsible for providing a 'correct' relation).