Add support for custom primary_key
Add a new optional parameter to the paginator to customize the primary
key field to use for a query. This defaults to `:id` if unset and also
affects ordering.

Setting a custom `primary_key` allows users to avoid always requesting
the `id` column in the query to compute the cursor. It is up to the user
to guarantee that the primary key is unique as this will affect ordering
as well.

The intended use case is for when paginating records that don't have an
`id` field or when paginating queries containing aggregations that are
not based on an `id` field.

Update Rubocop's rules to support the max length of the Paginator
class and the new number of parameters for the initializer.

Co-authored-by: Nicolas Fricke <1423331+nicolas-fricke@users.noreply.github.com>