Fix Gemfile for Postgres specs
While releasing the new 0.4.0 version of the gem, the
`Gemfile-postgres.lock` was not adjusted as required which caused the
test pipeline to fail.

This commit not only sets the gem version for the specs to 0.4.0 but
also applies all gem updates from the main `Gemfile.lock` to the
Postgres one.